Reduce Cyclomatic Complexity
* refactor: write context handling in MessageHandler and WriteWorker for improved clarity and reusability

* refactor: simplify message handling and improve batch processing in storage engine

* refactor: update request token handling and improve write resource management in storage engine

* refactor: enhance lock-free cache entry management and improve rate limiting logic in message handler

* refactor: add error handling for cache entry deinitialization in copy-on-write mutation

* refactor: replace entry deinitialization with allocator destroy for improved memory management

* refactor: simplify rate limiting calculation and improve batch dispatch logic in presence worker

* refactor: adjust rate limiting logic to use microsecond precision for burst capacity and token management

* refactor: replace allocator destroy with deinit for cache entry cleanup

+++ src/connection/state.zig
@@ -100,7 +100,7 @@ pub const Connection = struct {
     is_backpressured: bool = false,
     ref_count: std.atomic.Value(usize),
     created_at: i64,
-    request_tokens: f64,
+    request_tokens: u64,
     last_request_time: ?i64,
 
     pub fn initPool(self: *Connection, allocator: Allocator) void {
@@ -141,7 +141,7 @@ pub const Connection = struct {
         self.ws = ws;
         self.ref_count.store(1, .release);
         self.created_at = std.time.timestamp();
-        self.request_tokens = 0.0;
+        self.request_tokens = 0;
         self.last_request_time = null;
     }
